requirements.txt
anyio==3.4.0
asgiref==3.4.1
asyncpg==0.25.0
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
databases==0.5.3
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
iniconfig==1.1.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
psycopg2==2.9.2
py==1.11.0
pycodestyle==2.8.0
pydantic==1.8.2
pyparsing==3.0.6
pytest==6.2.5
python-dotenv==0.19.2
requests==2.27.1
sniffio==1.2.0
SQLAlchemy==1.4.28
starlette==0.16.0
toml==0.10.2
typing-extensions==4.0.1
urllib3==1.26.8
uvicorn==0.16.0
